DocuSorterAI/requirements.txt
Flask==2.3.3
spacy==3.7.2
scikit-learn==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.0.3
joblib==1.3.2
